break out params into modules
